+++ src/js/components/articles.js
@@ -5,7 +5,7 @@ import PaginationSearchHandler from './paginationSearchHandler.js';
 import { writeUserCards } from './dataBase/setDatabase';
 import { onGetCookie } from './dataBase/getCookie';
 import { getDatabase, ref, child, get } from 'firebase/database';
-export { findFavoriteCards, findReadCards };
+import { findFavoriteCards, findReadCards } from './findCardsInBase';
 
 const data = {
   source: 'nyt',
@@ -15,76 +15,9 @@ const data = {
 
 cleanLocalStorageFav();
 
-const refs = {
-  newsList: document.querySelector('.list-news'),
-};
+export const newsListRef = document.querySelector('.list-news');
 
-function findFavoriteCards() {
-  if (load('favCards')) {
-    const cardsArray = load('favCards');
-    const arrayHomePageCards = Array.from(
-      refs.newsList.querySelectorAll('.item-news__article')
-    );
-
-    arrayHomePageCards.forEach(card => {
-      const cardBtn = card.querySelector('.item-news__add-text');
-      const cardHeartImg = card.querySelector('.item-news__icon');
-      const cardTitle = card.querySelector('.item-news__title');
-      cardsArray.forEach(({ title }) => {
-        if (title === cardTitle.textContent.trim()) {
-          cardBtn.textContent = 'Remove from favorite';
-          cardHeartImg.classList.add('is-saved');
-        }
-      });
-    });
-  }
-}
-
-function findReadCards() {
-  if (onGetCookie('user')) {
-    const userId = onGetCookie('user');
-    findReadDataBase(userId);
-  } else {
-    if (load('readCards')) {
-      const cardsArray = load('readCards');
-      findReadLocalStorage(cardsArray);
-    }
-  }
-}
-
-function findReadLocalStorage(array) {
-  const arrayHomePageCards = Array.from(
-    refs.newsList.querySelectorAll('.item-news__article')
-  );
-
-  arrayHomePageCards.forEach(card => {
-    const cardStatus = card.querySelector('.item-news__already-read');
-    const cardTitle = card.querySelector('.item-news__title');
-    array.forEach(({ title }) => {
-      if (title === cardTitle.textContent.trim()) {
-        cardStatus.classList.add('is-read');
-        card.classList.add('is-ghost');
-      }
-    });
-  });
-}
-
-function findReadDataBase(userId) {
-  const dbRef = ref(getDatabase());
-
-  get(child(dbRef, `users/${userId}/readCards`))
-    .then(snapshot => {
-      if (snapshot.exists()) {
-        const cardsObject = snapshot.val();
-        findReadLocalStorage(cardsObject);
-      }
-    })
-    .catch(error => {
-      console.error(error);
-    });
-}
-
-refs.newsList.addEventListener('click', handleClickGallery);
+newsListRef.addEventListener('click', handleClickGallery);
 
 function handleClickGallery(e) {
   e.preventDefault();
